Hold position on the current bug displayed if the user is scrolling through the bugs.
This is to avoid a problem when a rapidly-occurring bug forces a complete refresh and forces the display position back to the end of the bugs list repeatedly.
A rapid-bug prevented a user from scrolling through the bugs to find the first bug which broke an addon, especially if there are many bugs to scroll through and the rapid-bug occurs in a loop/OnUpdate for example.
That made it extremely difficult for a developer or a user to identify the real cause of bugs in an addon, to get support and get a rapid bug fix.
This resolves Issue #68 Rapid Bugs Prevent Scroll Back to Previous Bugs.